% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_parquet.R
\name{csv_to_parquet}
\alias{csv_to_parquet}
\title{Convert a csv table to parquet format}
\usage{
csv_to_parquet(
  path_to_csv,
  url_to_csv,
  csv_as_a_zip = FALSE,
  filename_in_zip,
  path_to_parquet
)
}
\arguments{
\item{path_to_csv}{string that indicates the path to the csv file}

\item{url_to_csv}{string that indicates the URL of the csv file}

\item{csv_as_a_zip}{boolean that indicates if the csv is stored in a zip}

\item{filename_in_zip}{name of the csv file in the zip (useful if several csv are included in the zip). Required if `csv_as_a_zip` is TRUE.}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet file will be stored}
}
\value{
A parquet file
}
\description{
This function allows to convert a csv table to parquet format. \cr
}
\details{
Several conversion possibilities are offered :

\itemize{

\item{From a locally stored file. The argument `path_to_csv` must then be used;}
\item{From a URL. The argument `url_to_csv` must then be used.}

}
}
\examples{

\dontrun{
# Conversion from a local csv file :

csv_to_parquet(
  path_to_csv = "Data/ac1.csv",
  path_to_parquet = "Data",
)

# Conversion frome a URL and a csv file :

csv_to_parquet(
  url_to_csv = "https://stats.govt.nz/assets/Uploads/Research-and-development-survey/Research-and-development-survey-2021/Download-data/research-and-development-survey-2021-csv.csv",
  path_to_parquet = "Data",
)

# Conversion from a URL and a zipped file :

csv_to_parquet(
  url_to_csv = "https://www.insee.fr/fr/statistiques/fichier/3568617/equip-tour-transp-infra-2021.zip",
  csv_as_a_zip = TRUE,
  filename_in_zip = "equip-tour-transp-infra-2021.csv",
  path_to_parquet = "Data",
)
}
}
