% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_extract.R
\name{download_extract}
\alias{download_extract}
\title{download file and uncompress if needed}
\usage{
download_extract(path, filename_in_zip)
}
\arguments{
\item{path}{the input file}

\item{filename_in_zip}{name of the csv file in the zip. Required if several csv are included in the zip.}
}
\value{
the name of a file
}
\description{
This method will download the file or use the local file
}
\details{
if the file is a zip it :

* returns the file if zip is only one file
* returns filename_in_zip if there's more than one file

If the file is not a zip it returns immediately the path argument.
}
\examples{

download_extract(system.file("extdata","mtcars.csv.zip", package = "readr"))

download_extract(
  "https://www.nomisweb.co.uk/output/census/2021/census2021-ts007.zip",
  filename_in_zip = "census2021-ts007-ctry.csv"
)

download_extract("/my/local/file.csv")

}
