% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{expect_missing_argument}
\alias{expect_missing_argument}
\title{Check if missing argument error is raised}
\usage{
expect_missing_argument(object, regexp)
}
\arguments{
\item{object}{the object to check}

\item{regexp}{a regexp with the message we must find}
}
\value{
same as expect_error
}
\description{
Check if missing argument error is raised
}
\keyword{internal}
