% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{expect_parquet}
\alias{expect_parquet}
\title{Check if parquet dataset/file is readable and has the good number of rows}
\usage{
expect_parquet(
  path,
  with_lines,
  with_partitions = NULL,
  with_columns = NULL,
  with_files = NULL
)
}
\arguments{
\item{path}{to the parquet file or dataset}

\item{with_lines}{number of lines the file/dataset should have}

\item{with_partitions}{NULL or a vector with the partition names the dataset should have}

\item{with_columns}{NULL or a column's name vector the dataset/file should have}

\item{with_files}{NULL or number of files a dataset should have}
}
\value{
the dataset handle
}
\description{
Check if parquet dataset/file is readable and has the good number of rows
}
\keyword{internal}
