% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parquet_info.R
\name{get_parquet_info}
\alias{get_parquet_info}
\title{Get mean row group size of parquet files}
\usage{
get_parquet_info(path)
}
\arguments{
\item{path}{parquet file path or dataset}
}
\value{
a tibble with 4 columns :
\itemize{
\item path, the file path
\item num_rows, the number of rows
\item num_row_groups,  the number of groups
\item row_group_size, the mean nb of rows per group (row group size)
}

If the three last are \code{NA}, parquet file is malformed
}
\description{
Get mean row group size of parquet files
}
\examples{
get_parquet_info(system.file("extdata",package = "parquetize"))
}
