% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json_to_parquet.R
\name{json_to_parquet}
\alias{json_to_parquet}
\title{Convert a json file to parquet format}
\usage{
json_to_parquet(
  path_to_json,
  path_to_parquet,
  format = "json",
  progressbar = "yes",
  ...
)
}
\arguments{
\item{path_to_json}{string that indicates the path to the csv file}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet file will be stored}

\item{format}{string that indicates if the format is "json" (by default) or "ndjson"}

\item{progressbar}{string () ("yes" or "no" - by default) that indicates whether you want a progress bar to display}

\item{...}{additional format-specific arguments, see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}
and \href{https://arrow.apache.org/docs/r/reference/write_dataset.html}{arrow::write_dataset()} for more informations.}
}
\value{
A parquet file, invisibly
}
\description{
This function allows to convert a json file to parquet format. \cr
}
\details{
Several conversion possibilities are offered :

\itemize{

\item{From a locally stored file. Argument `path_to_csv` must then be used;}
\item{From a URL. Argument `url_to_csv` must then be used.}

}
}
\examples{

# Conversion from a local json file :

json_to_parquet(
  path_to_json = system.file("extdata","iris.json",package = "parquetize"),
  path_to_parquet = tempdir(),
  progressbar = "no"
)

# Conversion from a local ndjson file :

json_to_parquet(
  path_to_json = system.file("extdata","iris.ndjson",package = "parquetize"),
  path_to_parquet = tempdir(),
  format = "ndjson",
  progressbar = "no"
)
}
