% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-parquetize.R
\docType{package}
\name{parquetize-package}
\alias{parquetize}
\alias{parquetize-package}
\title{parquetize: Convert Files to Parquet Format}
\description{
Collection of functions to get files in parquet format. Parquet is a columnar storage file format \url{https://parquet.apache.org/}. The files to convert can be of several formats ("csv", "RData", "rds", "RSQLite", "json", "ndjson", "SAS", "SPSS"...).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ddotta.github.io/parquetize/}
  \item \url{https://github.com/ddotta/parquetize}
  \item Report bugs at \url{https://github.com/ddotta/parquetize/issues}
}

}
\author{
\strong{Maintainer}: Damien Dotta \email{damien.dotta@live.fr}

Authors:
\itemize{
  \item Nicolas Chuche \email{nicolas.chuche@barna.be}
}

}
\keyword{internal}
