% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbind_parquet.R
\name{rbind_parquet}
\alias{rbind_parquet}
\title{Function to bind multiple parquet files by row}
\usage{
rbind_parquet(folder, output_name, delete_initial_files = TRUE)
}
\arguments{
\item{folder}{the folder where the initial files are stored}

\item{output_name}{name of the output parquet file}

\item{delete_initial_files}{Boolean. Should the function delete the initial files ? By default TRUE.}
}
\value{
Parquet files, invisibly
}
\description{
This function read all parquet files in `folder` argument that starts with `output_name`,
combine them using rbind and write the result to a new parquet file. \cr
}
\details{
It can also delete the initial files if `delete_initial_files` argument is TRUE. \cr

Be careful, this function will not work if files with different structures
are present in the folder given with the argument `folder`.
}
\examples{
if (file.exists('output')==FALSE) {
  dir.create("output")
}

file.create(fileext = "output/parquetize0.parquet")
write_parquet(data.frame(
  x = c("a","b","c"),
  y = c(1L,2L,3L)
),

file.create(fileext = "output/parquetize500000.parquet")
write_parquet(data.frame(
  x = c("d","e","f"),
  y = c(4L,5L,6L)
), "output/parquetize500000.parquet")
test_data <- rbind_parquet(folder = "output", output_name = "test_data.parquet")
}
