% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rds_to_parquet.R
\name{rds_to_parquet}
\alias{rds_to_parquet}
\title{Convert a rds file to parquet format}
\usage{
rds_to_parquet(
  path_to_rds,
  path_to_parquet,
  format = "rds",
  partition = "no",
  progressbar = "yes",
  ...
)
}
\arguments{
\item{path_to_rds}{string that indicates the path to the csv file}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet file will be stored}

\item{partition}{string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file.
If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.}

\item{progressbar}{string () ("yes" or "no" - by default) that indicates whether you want a progress bar to display}

\item{...}{additional format-specific arguments, see \href{https://arrow.apache.org/docs/r/reference/write_parquet.html}{arrow::write_parquet()}
and \href{https://arrow.apache.org/docs/r/reference/write_dataset.html}{arrow::write_dataset()} for more informations.}
}
\value{
A parquet file, invisibly
}
\description{
This function allows to convert a rds file to parquet format. \cr
}
\details{
Two conversions possibilities are offered :

\itemize{

\item{Convert to a single parquet file. Argument `path_to_parquet` must then be used;}
\item{Convert to a partitioned parquet file. Additionnal arguments `partition` and `partitioning` must then be used;}

}
}
\examples{

# Conversion from a local rds file :

rds_to_parquet(
  path_to_rds = system.file("extdata","iris.rds",package = "parquetize"),
  path_to_parquet = tempdir(),
  progressbar = "no"
)

# Conversion from a local ndrds file :

rds_to_parquet(
  path_to_rds = system.file("extdata","iris.ndrds",package = "parquetize"),
  path_to_parquet = tempdir(),
  format = "ndrds",
  progressbar = "no"
)
}
