% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi_to_parquet.R
\name{stringify_sql}
\alias{stringify_sql}
\title{Convert a SQL query string in a filename's string}
\usage{
stringify_sql(sql_query)
}
\arguments{
\item{sql_query}{}
}
\value{
a string suitable to be a file name
}
\description{
This function take a SQL query string and replace all chararacters but a-z, A-Z and 0-9 by "_"
}
\keyword{internal}
