% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_parquet.R
\name{table_to_parquet}
\alias{table_to_parquet}
\title{Convert a csv table to parquet format}
\usage{
table_to_parquet(path_to_table, path_to_parquet, partition, ...)
}
\arguments{
\item{path_to_table}{string that indicates the path to the csv file}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet file will be stored}

\item{partition}{string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file}
}
\value{
Parquet files, invisibly
}
\description{
This function allows to convert a csv table to parquet format. \cr
}
\details{
Several conversion possibilities are offered :

\itemize{

\item{From a locally stored file. The argument `path_to_csv` must then be used;}
\item{From a URL. The argument `url_to_csv` must then be used.}

}
}
\examples{

\dontrun{
# Conversion from a local csv file :

table_to_parquet(
  path_to_csv = "Data/ac1.csv",
  path_to_parquet = "Data",
)

# Conversion from a URL and a csv file :

table_to_parquet(
  url_to_csv = "https://stats.govt.nz/assets/Uploads/Research-and-development-survey/Research-and-development-survey-2021/Download-data/research-and-development-survey-2021-csv.csv",
  path_to_parquet = "Data",
  compression = "gzip"
)

# Conversion from a URL and a zipped file :

table_to_parquet(
  url_to_csv = "https://www.insee.fr/fr/statistiques/fichier/3568617/equip-tour-transp-infra-2021.zip",
  csv_as_a_zip = TRUE,
  filename_in_zip = "equip-tour-transp-infra-2021.csv",
  path_to_parquet = "Data",
)
}
}
