% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_parquet.R
\name{csv_to_parquet}
\alias{csv_to_parquet}
\title{Convert a csv table to parquet format}
\usage{
csv_to_parquet(
  path_to_csv,
  url_to_csv,
  csv_as_a_zip = FALSE,
  filename_in_zip,
  path_to_parquet,
  compression = "snappy",
  compression_level = NULL,
  encoding = "UTF-8",
  progressbar = "yes",
  ...
)
}
\arguments{
\item{path_to_csv}{string that indicates the path to the csv file}

\item{url_to_csv}{string that indicates the URL of the csv file}

\item{csv_as_a_zip}{boolean that indicates if the csv is stored in a zip}

\item{filename_in_zip}{name of the csv file in the zip (useful if several csv are included in the zip). Required if `csv_as_a_zip` is TRUE.}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet file will be stored}

\item{compression}{compression algorithm. Default "snappy".}

\item{compression_level}{compression level. Meaning depends on compression algorithm.}

\item{encoding}{string that indicates the character encoding for the input file.}

\item{progressbar}{string () ("yes" or "no" - by default) that indicates whether you want a progress bar to display}

\item{...}{additional format-specific arguments, see [arrow::write_parquet()] and [arrow::write_dataset()] for more informations.}
}
\value{
A parquet file, invisibly
}
\description{
This function allows to convert a csv table to parquet format. \cr
}
\details{
Several conversion possibilities are offered :

\itemize{

\item{From a locally stored file. Argument `path_to_csv` must then be used;}
\item{From a URL. Argument `url_to_csv` must then be used.}

}
}
\examples{

# Conversion from a local csv file :

csv_to_parquet(
  path_to_csv = parquetize_example("region_2022.csv"),
  path_to_parquet = tempdir()
)

# Conversion from a URL and a csv file with "gzip" compression :

csv_to_parquet(
  url_to_csv = "https://stats.govt.nz/assets/Uploads/Research-and-development-survey/Research-and-development-survey-2021/Download-data/research-and-development-survey-2021-csv.csv",
  path_to_parquet = tempdir(),
  compression = "gzip",
  compression_level = 5
)

# Conversion from a URL and a zipped file :

csv_to_parquet(
  url_to_csv = "https://www.insee.fr/fr/statistiques/fichier/5057840/cog_ensemble_2021_csv.zip",
  csv_as_a_zip = TRUE,
  filename_in_zip = "commune2021.csv",
  path_to_parquet = tempdir()
)
}
