% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testthat-helpers.R
\name{expect_parquet}
\alias{expect_parquet}
\title{Check if parquet dataset/file is readable and has the good number of rows}
\usage{
expect_parquet(path, with_lines)
}
\arguments{
\item{path}{to the parquet file or dataset}

\item{with_lines}{number of lines we should have}
}
\value{
the dataset handle
}
\description{
Check if parquet dataset/file is readable and has the good number of rows
}
\keyword{internal}
