% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parquetize_example.R
\name{parquetize_example}
\alias{parquetize_example}
\title{Get path to parquetize example}
\usage{
parquetize_example(file = NULL)
}
\arguments{
\item{file}{Name of file. If `NULL`, the example files will be listed.}
}
\value{
A character string
}
\description{
parquetize comes bundled with a number of sample files in its `inst/extdata`
directory. This function make them easy to access
}
\examples{
parquetize_example()
parquetize_example("region_2022.csv")
parquetize_example("iris_dataset")
}
