% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_parquet.R
\name{table_to_parquet}
\alias{table_to_parquet}
\title{Convert an input table to parquet format}
\usage{
table_to_parquet(
  path_to_table,
  path_to_parquet,
  nb_rows = NULL,
  partition = "no",
  encoding = NULL,
  ...
)
}
\arguments{
\item{path_to_table}{string that indicates the path to the input file (don't forget the extension).}

\item{path_to_parquet}{string that indicates the path to the directory where the parquet files will be stored.}

\item{nb_rows}{By default NULL. Number of rows to process at once. This is the number of lines put into R's RAM and the number of lines written to disk for the parquet file.}

\item{partition}{string ("yes" or "no" - by default) that indicates whether you want to create a partitioned parquet file.
If "yes", `"partitioning"` argument must be filled in. In this case, a folder will be created for each modality of the variable filled in `"partitioning"`.}

\item{encoding}{string that indicates the character encoding for the input file.}

\item{...}{additional format-specific arguments, see [arrow::write_parquet()] and [arrow::write_dataset()] for more informations.}
}
\value{
Parquet files, invisibly
}
\description{
This function allows to convert an input table to parquet format. \cr
}
\details{
Several conversion possibilities are offered :

\itemize{

\item{From a locally stored file. The argument `path_to_csv` must then be used;}
\item{From a URL. The argument `url_to_csv` must then be used.}

}
}
\examples{

\dontrun{
# Conversion from a SAS file to a single parquet file :

table_to_parquet(
  path_to_table = "Data/postp.sas7bdat",
  path_to_parquet = "Data",
)

# Conversion from a SAS file to a single parquet file :

table_to_parquet(
  path_to_table = "Data/postp.sas7bdat",
  path_to_parquet = "Data",
)

# Conversion from a SPSS file to a single parquet file :

table_to_parquet(
  path_to_table = "Data/postp.sav",
  path_to_parquet = "Data"
)

# Conversion from a Stata file to a single parquet file :

table_to_parquet(
  path_to_table = "Data/postp.dta",
  path_to_parquet = "Data",
)

# Conversion from a big SAS file to a single parquet file by chunk :

table_to_parquet(
  path_to_table = "Data/postp.sas7bdat",
  path_to_parquet = "Data",
  nb_rows = 100000
)

# Conversion from a SAS file to a partitioned parquet file :

table_to_parquet(
  path_to_table = "Data/postp.sas7bdat",
  path_to_parquet = "Data",
  partition = "yes",
  partitioning =  c("REG")
)
}
}
